+++ doc/version_history.rst
@@ -12,6 +12,7 @@ v1.21.0
 * Add new ``maintel/m1m3/lower_m1m3.py`` sal script and associated files. 
 * Add new ``auxtel/offset_ataos.py`` script to offset the ATAOS.
 * Add new ``maintel/m1m3/check_hardpoint.py`` script to check hardpoints.
+* Add missing comment line in all script files. 
 
 v1.20.1
 -------