+++ doc/version_history.rst
@@ -6,6 +6,18 @@
 Version History
 ===============
 
+v1.21.0
+-------
+
+* Add new ``maintel/m1m3/check_actuators.py`` script to run the actuators bump test.
+* Add new ``maintel/m1m3/lower_m1m3.py`` sal script and associated files. 
+* Add new ``auxtel/offset_ataos.py`` script to offset the ATAOS.
+* Add new ``maintel/m1m3/check_hardpoint.py`` script to check hardpoints.
+* Add missing comment line in all script files. 
+* In ``auxtel/offset_ataos.py``, fix bug in call to resetOffset and change handling for reset all configuration.
+* Update unit test for ``auxtel/offset_ataos.py``
+* In ``auxtel/daytime_checkout/atpneumatics_checkout.py``, update detailed description. 
+
 v1.20.1
 -------
 
@@ -13,7 +25,6 @@ v1.20.1
 * Update ``auxtel/latiss_take_sequence.py`` to configure synchronization between ``ATCS`` and ``LATISS``.
 * Update ts-pre-commit configuration.
 * In ``base_offset_tcs.py``:
-
   * Add new option to execute ``offset_pa``.
   * Add checkpoints for each action.